Major pawn spawn change
This is a breaking change. Pawns no longer need to be registered, and Actors now define a getter to return the class of their pawn instead of it being passed to init.

More cleanup needs to happen and I need to verity that all Worldcore examples are converted to this new scheme.